chore!: Drop Ruby 2.5 and earlier support
To get the latest ver of commonmarker, which has security vuln fixes,
this project needs to drop support for older Ruby versions. As of this
commit, the stable release series of Ruby are:

- 3.1
- 3.0
- 2.7

Ruby 2.6 is EOL but it seems to be working still, so this continues to
maintain support for it.