feat: replace Ruby sass with dartsass
The Ruby sass gem has been deprecated and is no longer actively maintained.
This replaces it with the dartsass gem, which uses a compiled ver of the
canonical sass compiler. It allows for `bundle exec dartsass` compile
styles. It has slightly different CLI options.

I'm truthfully on the fence about whether or not this project even needs
Sass, but it's easier to drop this in and maintain the existing
styles than to combine them all and deal with a 1k+ line stylesheet.

Fixes https://github.com/brettchalupa/graphql-docs/issues/86

BREAKING CHANGE: removes sass gem; if you rely on it, bundle it directly or call out to dartsass