Add support for running as a Rack app
This makes it so that users of the gem can optionally run the app as a
Rack app, making it easy to mount within a Rails or Sinatra app. Reasons
why would be to add authentication or have a nicer developer experience
instead of regenerating the docs over and over again.

Fixes https://github.com/brettchalupa/graphql-docs/issues/102